Separate User List Page's list of users into a new component
