+++ front-end/src/pages/userpages/userlistpage/UserListPage.css
@@ -4,25 +4,8 @@
     text-align: center;
 }
 
-.numUsers {
-    font-size: 15px;
-    text-align: center;
-    margin-top: -4px;
-    margin-bottom: 20px;
-}
-
 .backLink {
     font-size: 26px;
     margin-top: -10px;
     position: absolute;
-}
-
-.userSearchBar {
-    margin-bottom: 30px;
-}
-
-.noUsersFoundMessage {
-    font-size: 14px;
-    font-weight: lighter;
-    text-align: center;
 }
